Update integration test cases and helpers to work within site sessions
